@@ -1,12 +1,10 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk.Web">
   <PropertyGroup>
-    <TargetFramework>netcoreapp2.2</TargetFramework>
-    <ReleaseVersion>1.9</ReleaseVersion>
+    <TargetFramework>netcoreapp3.1</TargetFramework>
+    <IncludeOpenAPIAnalyzers>true</IncludeOpenAPIAnalyzers>
+    <ReleaseVersion>1.11</ReleaseVersion>
     <PublishWithAspNetCoreTargetManifest>false</PublishWithAspNetCoreTargetManifest>
     <AspNetCoreHostingModel>InProcess</AspNetCoreHostingModel>
-  </PropertyGroup>
-  <PropertyGroup>
-    <CodeAnalysisRuleSet>JewishBot.ruleset</CodeAnalysisRuleSet>
   </PropertyGroup>
   <ItemGroup>
     <Folder Include="Controllers\" />
@@ -19,20 +17,14 @@
   </ItemGroup>
   <ItemGroup>
     <PackageReference Include="CsQuery.NETStandard" Version="1.3.6.1" />
-    <PackageReference Include="Microsoft.CodeAnalysis.FxCopAnalyzers" Version="2.6.2" />
     <PackageReference Include="Api.Forex.Currency.Converter" Version="1.0.5" />
-    <PackageReference Include="Microsoft.AspNetCore" Version="2.2.0" />
-    <PackageReference Include="Microsoft.AspNetCore.Mvc" Version="2.2.0" />
-    <PackageReference Include="GeoTimeZone" Version="3.2.0" />
-    <PackageReference Include="Telegram.Bot" Version="14.10.0" />
-    <PackageReference Include="Microsoft.EntityFrameworkCore.Sqlite" Version="2.2.0" />
-    <PackageReference Include="Microsoft.Extensions.Http" Version="2.2.0" />
+    <PackageReference Include="GeoTimeZone" Version="4.0.0" />
+    <PackageReference Include="Telegram.Bot" Version="15.1.0" />
+    <PackageReference Include="Microsoft.EntityFrameworkCore.Sqlite" Version="3.1.0" />
+    <PackageReference Include="Microsoft.AspNetCore.Mvc.NewtonsoftJson" Version="3.1.0" />
   </ItemGroup>
   <ItemGroup>
     <None Include="appsettings.json" CopyToPublishDirectory="Always" />
     <None Include="jewish_bot.pem" CopyToPublishDirectory="Always" />
-  </ItemGroup>
-  <ItemGroup>
-    <Compile Remove="Migrations\20180601132723_Initial.cs" />
   </ItemGroup>
 </Project>
